constrain templates to the age of the universe plus a 0.5 Gyr fudge factor
