+++ py/fastspecfit/continuum.py
@@ -1081,7 +1081,7 @@ class ContinuumTools(object):
     
             if png:
                 fig.savefig(png)
-    
+
         else:
             linemask_dict = {'linemask_all': [], 'linemask': [],
                              'linename': [], 'linepix': [], 'contpix': []}
@@ -1284,3 +1284,185 @@ class ContinuumTools(object):
                     datatemplateflux = datatemplateflux.reshape(nwavepix, nsed, nprop) # [npix,nsed,nprop]
                 
         return datatemplateflux, templatephot # vector or 3-element list of [npix,nmodel] spectra
+
+    def get_mean_property(self, physical_property, coeff, agekeep,
+                          normalization=None, log10=False):
+        """Compute the mean physical properties, given a set of coefficients.
+
+        """
+        values = self.templateinfo[physical_property][agekeep] # account for age of the universe trimming
+
+        if np.count_nonzero(coeff > 0) == 0:
+            self.log.warning('Coefficients are all zero!')
+            meanvalue = 0.0
+            #raise ValueError
+        else:
+            meanvalue = values.dot(coeff)
+            # the coefficients include the stellar mass normalization
+            if physical_property != 'mstar' and physical_property != 'sfr':
+                meanvalue /= np.sum(coeff) 
+            if normalization:
+                meanvalue /= normalization
+            if log10 and meanvalue > 0:
+                meanvalue = np.log10(meanvalue)
+        
+        return meanvalue
+
+    def younger_than_universe(self, redshift, agepad=0.5):
+        """Return the indices of the templates younger than the age of the universe
+        (plus an agepadding amount) at the given redshift.
+
+        agepad in Gyr
+
+        """
+        return np.where(self.templateinfo['age'] <= (agepad*1e9 + self.cosmo.age(redshift).to(u.year).value))[0]
+
+    def kcorr_and_absmag(self, data, continuum, coeff, snrmin=2.0):
+        """Computer K-corrections, absolute magnitudes, and a simple stellar mass.
+
+        """
+        from scipy.stats import sigmaclip
+        from scipy.ndimage import median_filter
+        
+        redshift = data['zredrock']
+        
+        if data['photsys'] == 'S':
+            filters_in = self.decamwise
+        else:
+            filters_in = self.bassmzlswise
+        lambda_in = filters_in.effective_wavelengths.value
+
+        # redshifted wavelength array and distance modulus
+        ztemplatewave = self.templatewave * (1 + redshift)
+        dmod = self.cosmo.distmod(redshift).value
+
+        maggies = data['phot']['nanomaggies'].data * 1e-9
+        ivarmaggies = (data['phot']['nanomaggies_ivar'].data / 1e-9**2) * self.bands_to_fit # mask W2-W4
+
+        # input bandpasses, observed frame; maggies and bestmaggies should be
+        # very close.
+        bestmaggies = filters_in.get_ab_maggies(continuum / self.fluxnorm, ztemplatewave)
+        bestmaggies = np.array(bestmaggies.as_array().tolist()[0])
+
+        # need to handle filters with band_shift!=0 separately from those with band_shift==0
+        def _kcorr_and_absmag(filters_out, band_shift):
+            nout = len(filters_out)
+
+            # note the factor of 1+band_shift
+            lambda_out = filters_out.effective_wavelengths.value / (1 + band_shift)
+
+            # Multiply by (1+z) to convert the best-fitting model to the "rest
+            # frame" and then divide by 1+band_shift to shift it and the
+            # wavelength vector to the band-shifted redshift. Also need one more
+            # factor of 1+band_shift in order maintain the AB mag normalization.
+            synth_outmaggies_rest = filters_out.get_ab_maggies(continuum * (1 + redshift) / (1 + band_shift) /
+                                                               self.fluxnorm, self.templatewave * (1 + band_shift))
+            synth_outmaggies_rest = np.array(synth_outmaggies_rest.as_array().tolist()[0]) / (1 + band_shift)
+    
+            # output bandpasses, observed frame
+            synth_outmaggies_obs = filters_out.get_ab_maggies(continuum / self.fluxnorm, ztemplatewave)
+            synth_outmaggies_obs = np.array(synth_outmaggies_obs.as_array().tolist()[0])
+    
+            absmag = np.zeros(nout, dtype='f4')
+            ivarabsmag = np.zeros(nout, dtype='f4')
+            kcorr = np.zeros(nout, dtype='f4')
+            for jj in np.arange(nout):
+                lambdadist = np.abs(lambda_in / (1 + redshift) - lambda_out[jj])
+                # K-correct from the nearest "good" bandpass (to minimizes the K-correction)
+                #oband = np.argmin(lambdadist)
+                #oband = np.argmin(lambdadist + (ivarmaggies == 0)*1e10)
+                oband = np.argmin(lambdadist + (maggies*np.sqrt(ivarmaggies) < snrmin)*1e10)
+                kcorr[jj] = + 2.5 * np.log10(synth_outmaggies_rest[jj] / bestmaggies[oband])
+
+                # m_R = M_Q + DM(z) + K_QR(z) or
+                # M_Q = m_R - DM(z) - K_QR(z)
+                if maggies[oband] * np.sqrt(ivarmaggies[oband]) > snrmin:
+                #if (maggies[oband] > 0) and (ivarmaggies[oband]) > 0:
+                    absmag[jj] = -2.5 * np.log10(maggies[oband]) - dmod - kcorr[jj]
+                    ivarabsmag[jj] = maggies[oband]**2 * ivarmaggies[oband] * (0.4 * np.log(10.))**2
+                else:
+                    # if we use synthesized photometry then ivarabsmag is zero
+                    # (which should never happen?)
+                    absmag[jj] = -2.5 * np.log10(synth_outmaggies_rest[jj]) - dmod
+
+                check = absmag[jj], -2.5*np.log10(synth_outmaggies_rest[jj]) - dmod
+                self.log.debug(check)
+
+            return kcorr, absmag, ivarabsmag
+
+        kcorr_01, absmag_01, ivarabsmag_01 = _kcorr_and_absmag(self.absmag_filters_01, band_shift=0.1)
+        kcorr_00, absmag_00, ivarabsmag_00 = _kcorr_and_absmag(self.absmag_filters_00, band_shift=0.0)
+
+        nout = len(self.absmag_bands)
+        kcorr = np.zeros(nout, dtype='f4')
+        absmag = np.zeros(nout, dtype='f4')
+        ivarabsmag = np.zeros(nout, dtype='f4')
+
+        I00 = np.isin(self.absmag_bands, self.absmag_bands_00)
+        I01 = np.isin(self.absmag_bands, self.absmag_bands_01)
+
+        kcorr[I00] = kcorr_00
+        absmag[I00] = absmag_00
+        ivarabsmag[I00] = ivarabsmag_00
+
+        kcorr[I01] = kcorr_01
+        absmag[I01] = absmag_01
+        ivarabsmag[I01] = ivarabsmag_01
+
+        nage = len(coeff)
+        
+        # From Taylor+11, eq 8
+        #mstar = self.templateinfo['mstar'][:nage].dot(coeff) * self.massnorm
+        #https://researchportal.port.ac.uk/ws/files/328938/MNRAS_2011_Taylor_1587_620.pdf
+        #mstar = 1.15 + 0.7*(absmag[1]-absmag[3]) - 0.4*absmag[3]
+
+        # compute the model continuum flux at 1500 and 2800 A (to facilitate UV
+        # luminosity-based SFRs) and at the positions of strong nebular emission
+        # lines [OII], Hbeta, [OIII], and Halpha
+        dfactor = (1 + redshift) * 4.0 * np.pi * self.cosmo.luminosity_distance(redshift).to(u.cm).value**2 / self.fluxnorm
+
+        lums = {}
+        cwaves = [1500.0, 2800.0, 5100.0]
+        labels = ['LOGLNU_1500', 'LOGLNU_2800', 'LOGL_5100']
+        norms = [1e28, 1e28, 1e10]
+        for cwave, norm, label in zip(cwaves, norms, labels):
+            J = (self.templatewave > cwave-500) * (self.templatewave < cwave+500)
+            I = (self.templatewave[J] > cwave-20) * (self.templatewave[J] < cwave+20)
+            smooth = median_filter(continuum[J], 200)
+            clipflux, _, _ = sigmaclip(smooth[I], low=1.5, high=3)
+            cflux = np.median(clipflux) # [flux in 10**-17 erg/s/cm2/A]
+            cflux *= dfactor # [monochromatic luminosity in erg/s/A]
+            if label == 'LOGL_5100':
+                cflux *= cwave / 3.846e33 / norm # [luminosity in 10**10 L_sun]
+            else:
+                # Convert the UV fluxes to rest-frame luminosity in erg/s/Hz. This
+                # luminosity can be converted into a SFR using, e.g., Kennicutt+98,
+                # SFR=1.4e-28 * L_UV
+                cflux *= cwave**2 / (C_LIGHT * 1e13) / norm # [monochromatic luminosity in 10**(-28) erg/s/Hz]
+            if cflux > 0:
+                lums[label] = np.log10(cflux) # * u.erg/(u.second*u.Hz)
+
+        cfluxes = {}
+        cwaves = [3728.483, 4862.683, 5008.239, 6564.613]
+        labels = ['FOII_3727_CONT', 'FHBETA_CONT', 'FOIII_5007_CONT', 'FHALPHA_CONT']
+        for cwave, label in zip(cwaves, labels):
+            J = (self.templatewave > cwave-500) * (self.templatewave < cwave+500)
+            I = (self.templatewave[J] > cwave-20) * (self.templatewave[J] < cwave+20)
+            smooth = median_filter(continuum[J], 200)
+            clipflux, _, _ = sigmaclip(smooth[I], low=1.5, high=3)
+            cflux = np.median(clipflux) # [flux in 10**-17 erg/s/cm2/A]
+            cfluxes[label] = cflux # * u.erg/(u.second*u.cm**2*u.Angstrom)
+            
+            #import matplotlib.pyplot as plt
+            #print(cwave, cflux)
+            #plt.clf()
+            #plt.plot(self.templatewave[J], continuum[J])
+            #plt.plot(self.templatewave[J], smooth, color='k')
+            #plt.axhline(y=cflux, color='red')
+            #plt.axvline(x=cwave, color='red')
+            #plt.xlim(cwave - 50, cwave + 50)
+            #plt.savefig('junk.png')
+            ##plt.savefig('desi-users/ioannis/tmp/junk.png')
+            #pdb.set_trace()
+
+        return kcorr, absmag, ivarabsmag, bestmaggies, lums, cfluxes
